// src/generated/client/types/PutRoutinesRequestExercise.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { PutRoutinesRequestSet } from "./PutRoutinesRequestSet.ts";

export type PutRoutinesRequestExercise = {
    /**
     * @description The ID of the exercise template.
     * @type string | undefined
    */
    exercise_template_id?: string;
    /**
     * @description The ID of the superset.
     * @type integer
    */
    superset_id?: number | null;
    /**
     * @description The rest time in seconds.
     * @type integer
    */
    rest_seconds?: number | null;
    /**
     * @description Additional notes for the exercise.
     * @type string
    */
    notes?: string | null;
    /**
     * @type array | undefined
    */
    sets?: PutRoutinesRequestSet[];
};