// src/generated/client/types/PutRoutinesRequestSet.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const putRoutinesRequestSetTypeEnum = {
    warmup: "warmup",
    normal: "normal",
    failure: "failure",
    dropset: "dropset"
} as const;

export type PutRoutinesRequestSetTypeEnumKey = (typeof putRoutinesRequestSetTypeEnum)[keyof typeof putRoutinesRequestSetTypeEnum];

export type PutRoutinesRequestSet = {
    /**
     * @description The type of the set.
     * @type string | undefined
    */
    type?: PutRoutinesRequestSetTypeEnumKey;
    /**
     * @description The weight in kilograms.
     * @type number
    */
    weight_kg?: number | null;
    /**
     * @description The number of repetitions.
     * @type integer
    */
    reps?: number | null;
    /**
     * @description The distance in meters.
     * @type integer
    */
    distance_meters?: number | null;
    /**
     * @description The duration in seconds.
     * @type integer
    */
    duration_seconds?: number | null;
    /**
     * @description A custom metric for the set. Currently used for steps and floors.
     * @type number
    */
    custom_metric?: number | null;
    /**
     * @description Range of reps for the set, if applicable
     * @type object
    */
    rep_range?: {
        /**
         * @description Starting rep count for the range
         * @type number
        */
        start?: number | null;
        /**
         * @description Ending rep count for the range
         * @type number
        */
        end?: number | null;
    } | null;
};